seyup

+++ install.sh
@@ -3,14 +3,11 @@
 # Aktualizacja repozytoriów
 sudo apt-get update
 
-# Instalacja python3-venv, jeśli nie jest zainstalowany
-sudo apt-get install -y python3-venv
+# Instalacja python3, jeśli nie jest zainstalowany
+sudo apt-get install -y python3-venv python3-dev
 
-sudo apt-get install pkg-config
-sudo apt-get install build-essential
-sudo apt-get install python3-dev
-sudo apt-get install libdouble-conversion-dev
-
+# Instalacja pakietów
+sudo apt-get install -y pkg-config build-essential libdouble-conversion-dev
 
 # Tworzenie wirtualnego środowiska
 python3 -m venv venv
@@ -18,9 +15,15 @@ python3 -m venv venv
 # Aktywacja wirtualnego środowiska
 source venv/bin/activate
 
-# Instalacja pakietów z requirements.txt
-pip install -r requirements.txt
+# Instalacja setuptools
+pip install setuptools
+
+# Instalacja pakietów z requirements.txt, jeśli istnieje
+if [ -f "requirements.txt" ]; then
+    pip install -r requirements.txt
+fi
+
+# Instalacja projektu za pomocą setuptools
+pip install .
 
 echo "Instalacja zakończona. Użyj 'source venv/bin/activate' aby aktywować wirtualne środowisko. Następnie uruchom 'python app.py' aby uruchomić program."
-
-
